refactor (auth) : resolve the some fixes
