+++ middleware/init.global.ts
@@ -1,20 +1,55 @@
 import { defineNuxtRouteMiddleware, navigateTo, useCookie } from "nuxt/app";
-import { userInfo } from "@/services/userService";
-import { useUser } from "~/composables/useUser";
+import { useUser } from "@/composables/useUser";
+import { useAuth } from "@/composables/useAuth";
+
+interface UserResponse {
+  data: any;
+}
+
+interface ErrorResponse {
+  response?: {
+    status: number;
+    data?: {
+      message: string;
+    };
+  };
+}
 
 export default defineNuxtRouteMiddleware(async (to) => {
+  // Skip for public routes
   const publicRoutes = ["/login", "/register", "/forgot-password"];
   if (publicRoutes.includes(to.path)) {
     return;
   }
-  const authToken = useCookie("authToken");
-  if (authToken.value) {
-    try {
-      const userData = await userInfo();
-      if (userData) {
-        const { setUser } = useUser();
-        setUser(userData);
+
+  const auth = useAuth();
+  const authToken = auth.getUserToken();
+
+  // Skip if no token
+  if (!authToken) {
+    if (to.path.startsWith("/user")) {
+      return navigateTo("/login");
+    }
+    return;
+  }
+
+  try {
+    const response = await $fetch<UserResponse>(
+      `/api/v1/users/info?uid=${authToken}`,
+      {
+        method: "GET",
       }
-    } catch (error) {}
+    );
+
+    if (response && response.data) {
+      const { setUser } = useUser();
+      setUser(response.data);
+    } else if (to.path.startsWith("/user")) {
+    }
+  } catch (error) {
+    const status = (error as ErrorResponse)?.response?.status;
+    if ((status === 401 || status === 403) && to.path.startsWith("/user")) {
+      // return navigateTo("/login");
+    }
   }
 });
